+++ Source/GateRimSG1/Goauld/HediffCompProperties_GoauldEmergencyExtraction.cs
@@ -1,0 +1,13 @@
+using Verse;
+
+namespace GateRimSG1.Goauld
+{
+    public class HediffCompProperties_GoauldEmergencyExtraction
+        : HediffCompProperties
+    {
+        public HediffCompProperties_GoauldEmergencyExtraction()
+        {
+            compClass = typeof(HediffComp_GoauldEmergencyExtraction);
+        }
+    }
+}